\name{EtiquetasGLNF}
\alias{EtiquetasGLNF}
\title{
Etiquetas de grados de libertad de los Nodos N y F.
}
\description{
Se colocan las etiquetas de los grados de libertad correspondiente a cada nodo. Además, respecto al nodo cercano (N)
y lejano (F).
}
\usage{
EtiquetasGLNF(EtiquetasGl, elementos, numElem, truss)
}
\arguments{
  \item{EtiquetasGl}{Etiquetas de los grados de libertad de cada nodo.}
  \item{elementos}{Elementos del sistema. Nodo a Nodo.}
  \item{numElem}{Número de elementos.}
  \item{truss}{Si las dimesiones son 2D, truss=2}
}
