\name{matrizReacciones}
\alias{matrizReacciones}

\title{
Mariz de reacciones
}
\description{
Se muestran las reacciones en los nodos y ejes correspondientes.
}
\usage{
matrizReacciones(Reacciones, restricciones, EtiquetasGl)
}
\arguments{
  \item{Reacciones}{Reacciones obtenidas con sus correspondientes grados de libertad.}
  \item{restricciones}{Nodos restringidos en su dirección.}
  \item{EtiquetasGl}{Etiquetas de grados de libertad de cada nodo.}
}
\details{
Se muestran las reacciones en los nodos y ejes correspondientes.
}
