\name{nodoscargas}
\alias{nodoscargas}
\title{
Nodo cargas
}
\description{
Nodos en donde irá el inicio y término de las flechas de cargas en 3D.
}
\usage{
nodoscargas(nodos, cargas, r, truss)
}
\arguments{
  \item{nodos}{Nodos del sistema.}
  \item{cargas}{Nodos con sus respectivas cargas.}
  \item{r}{longitud de la flecha de carga.}
  \item{truss}{Espacio 2 o 3D.}
}
\details{
Esta función se ocupa dentro de la función graficaModelo.R y 2D.
}
\value{
coordenadas en donde van las flechas de cargas.
}
