\name{nodosIntermedios}
\alias{nodosIntermedios}
\title{
Nodos intermedios para etiquetas del modelo.
}
\description{
Se calcula la mitad del elemento para colocar ahí las etiquetas de número de elemento.
}
\usage{
nodosIntermedios(elementos, nodos)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{elementos}{Elementos de la armadura.}
  \item{nodos}{Nodos de la armadura.}
}
