\name{reporte}
\alias{reporte}

\title{
Reporte del análisis
}
\description{
Resulados del análisis matricial.
}
\usage{
reporte(Resultados, imprimir = FALSE, uniLong = "metros", uniCarga = "Toneladas")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Resultados}{Resultados de aplicar la función análisis.R.}
  \item{imprimir}{Si es TRUE, devuélve la impresión de resultados en formato .txt.}
  \item{uniLong}{Unidad de longitud. Por default es "Metros".}
  \item{uniCarga}{Unidad de carga. Por Por default es "Toneladas".}
}
\details{
Reporte cn los resultados del análisis
}
